+++ app/context/TRQs/users/mutations/useAuthenticate.ts
@@ -2,26 +2,27 @@
 
 import useGlobalErrorHandler from "@/app/hooks/useGlobalErrorHandler";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { useRouter } from "next/navigation";
 import { authenticateAsync } from "pawpal-fe-common/users-api";
 import { IAuthenticatePayload } from "pawpal-fe-common/users-interfaces";
-import useAuthentication from "../../useAuthentication";
 import useCurrentUser from "../useCurrentUser";
+import { useAuth } from "@/app/context/AuthContext";
+import useListings from "../../listings/useListings";
 
 const useAuthenticate = () => {
   const queryClient = useQueryClient();
   const { handleError } = useGlobalErrorHandler();
-  const router = useRouter();
+  const { setAuthStatus } = useAuth();
 
   return useMutation({
     mutationFn: async (payload: IAuthenticatePayload) =>
       await authenticateAsync(payload),
     onSuccess: () => {
-      queryClient.invalidateQueries({
-        queryKey: [useAuthentication.name],
-      });
+      setAuthStatus(true);
       queryClient.invalidateQueries({
         queryKey: [useCurrentUser.name],
+      });
+      queryClient.invalidateQueries({
+        queryKey: [useListings.name],
       });
     },
     onError: (error) => {