+++ app/context/TRQs/users/mutations/useSignOutBE.ts
@@ -1,0 +1,15 @@
+"use client";
+
+import { useMutation } from "@tanstack/react-query";
+import { signOutAsync } from "pawpal-fe-common/users-api";
+
+const useSignOutBE = () => {
+  return useMutation({
+    mutationFn: async () => signOutAsync(),
+    onError: (error) => {
+      // Dismiss
+    },
+  });
+};
+
+export default useSignOutBE;